requirements.txt
bleach>=2.0.0
coverage>=4.4.1
dj-database-url>=0.4.2
dj-static>=0.0.6
django-autoslug>=1.9
django-taggit>=0.22
Django>=1.10
mysqlclient>=1.3.6
gunicorn>=19.7
Markdown>=2.6
Pillow>=4.2
psycopg2>=2.7
python-decouple>=3
pytz>=2017.2
Unipath>=1.1
whitenoise>=3.3.0
